MAILBOX-258 Relocate ReIndexers in store to avoid circular dependencies
