+++ mailbox/store/src/main/java/org/apache/james/mailbox/store/search/indexer/events/ImpactingEvent.java
@@ -17,31 +17,14 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.indexer.registrations;
+package org.apache.james.mailbox.store.search.indexer.events;
 
-import org.apache.james.mailbox.MailboxListener;
 import org.apache.james.mailbox.model.MailboxPath;
 
-import java.util.concurrent.ConcurrentHashMap;
+public interface ImpactingEvent {
 
-public class GlobalRegistration implements MailboxListener {
+    MailboxPath getMailboxPath();
 
-    private final ConcurrentHashMap<MailboxPath, Boolean> impactingEvents;
+    ImpactingEventType getType();
 
-    public GlobalRegistration() {
-        this.impactingEvents = new ConcurrentHashMap<MailboxPath, Boolean>();
-    }
-
-    public boolean indexThisPath(MailboxPath mailboxPath) {
-        return impactingEvents.get(mailboxPath) != null;
-    }
-
-    @Override
-    public void event(Event event) {
-        if (event instanceof MailboxDeletion) {
-            impactingEvents.putIfAbsent(event.getMailboxPath(), true);
-        } else if (event instanceof Expunged) {
-            impactingEvents.putIfAbsent(event.getMailboxPath(), true);
-        }
-    }
 }
