+++ mailbox/store/src/main/java/org/apache/james/mailbox/store/search/indexer/events/ImpactingMessageEvent.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.indexer.events;
+package org.apache.james.mailbox.store.search.indexer.events;
 
 public interface ImpactingMessageEvent extends ImpactingEvent {
