+++ server/container/mailbox-adapter/pom.xml
@@ -59,10 +59,6 @@
             <artifactId>apache-james-mailbox-maildir</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.apache.james</groupId>
-            <artifactId>apache-james-mailbox-tool</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
         </dependency>
